+++ src/MQTTAnalyzer/model/v2/TreeModel.swift
@@ -7,81 +7,6 @@
 //
 
 import Foundation
-import SwiftyJSON
-
-class MsgPayload {
-	let data: [UInt8]
-	var jsonData: JSON?
-	var isJSON: Bool {
-		return jsonData != nil
-	}
-	var isBinary: Bool {
-		return dataStringCache == nil
-	}
-	
-	private let dataStringCache: String?
-	var dataString: String {
-		return dataStringCache ?? "[\(data.count) bytes]"
-	}
-
-	init(data: [UInt8]) {
-		self.data = data
-		self.dataStringCache = MsgPayload.toOptionalString(data: data)
-		self.jsonData = MsgPayload.toJson(str: dataStringCache)
-	}
-}
-
-extension MsgPayload {
-	var prettyJSON: String {
-		return dataStringCache != nil ? JSONUtils.format(json: dataStringCache!) : ""
-	}
-}
-
-extension MsgPayload {
-	class func toJson(str: String?) -> JSON? {
-		if str == nil {
-			return nil
-		}
-		
-		let json = JSON.init(parseJSON: str!)
-		if json.isEmpty {
-			return nil
-		}
-		else {
-			return json
-		}
-	}
-	
-	class func toOptionalString(data: [UInt8]) -> String? {
-		return NSString(bytes: data, length: data.count, encoding: String.Encoding.utf8.rawValue) as String?
-	}
-}
-
-class MsgMetadata {
-	let date: Date
-	let localDate: String
-	let qos: Int32
-	let retain: Bool
-	
-	init(qos: Int32, retain: Bool) {
-		self.date = Date.now
-		self.localDate = DateFormatter.iso.string(from: self.date)
-		self.qos = qos
-		self.retain = retain
-	}
-}
-
-class MsgMessage: Identifiable {
-	let topic: TopicTree
-	let payload: MsgPayload
-	let metadata: MsgMetadata
-	
-	init(topic: TopicTree, payload: MsgPayload, metadata: MsgMetadata) {
-		self.topic = topic
-		self.payload = payload
-		self.metadata = metadata
-	}
-}
 
 class TopicTree: Identifiable, ObservableObject {
 	let id = UUID()
